refactor: product image 분리
